+++ modules/widgets/src/loading-widget.tsx
@@ -5,7 +5,7 @@
 import type {WidgetPlacement, Layer} from '@deck.gl/core';
 import {render} from 'preact';
 import {Widget, WidgetProps} from '@deck.gl/core';
-import {IconButton} from './lib/components';
+import {IconButton} from './lib/components/icon-button';
 
 /** Properties for the LoadingWidget */
 export type LoadingWidgetProps = WidgetProps & {
