+++ modules/widgets/src/screenshot-widget.tsx
@@ -6,7 +6,7 @@
 import type {WidgetPlacement} from '@deck.gl/core';
 import {render} from 'preact';
 import {Widget, WidgetProps} from '@deck.gl/core';
-import {IconButton} from './lib/components';
+import {IconButton} from './lib/components/icon-button';
 
 /** Properties for the ScreenshotWidget */
 export type ScreenshotWidgetProps = WidgetProps & {
