+++ modules/widgets/src/view-selector-widget.tsx
@@ -5,7 +5,7 @@
 
 import {render, JSX} from 'preact';
 import {Widget, type WidgetProps, type WidgetPlacement} from '@deck.gl/core';
-import {IconMenu} from './lib/icon-menu';
+import {IconMenu} from './lib/components/icon-menu';
 
 /** The available view modes */
 export type ViewMode = 'single' | 'split-horizontal' | 'split-vertical';
